Add ClaimReview, improve preset descriptions
* Added basic ClaimReview check
* Added better descriptions of schemas from Schema.org

The ClaimReview check is a good example of where conditional checks might be useful.

e.g. if a proprety is not specified, genreate a warning - but if it is specified and missing required fields generate an error.

This is actually a confusing aspect of the Google Stuctured Data Testing Tool - it will flag some fields as 'warnings' but then be very strict about what sort of input is isable to handle and generate errors if you use them in ways it does not support or if the object is specified by missing additional properties it expect it then flags 'errors' (suggesting it is better to specify nothing at all for them).