Improve output of --info option
Shows values found for tests when using `--info `option on the CLI.

Value returned for a test is also added to `value` property of a test when using the API.

Has specific handling to nicely format when there are multiple results for a test.

Improved output of `--schema` option via CLI as part of this work.