Add support for JSON input
Resolves #9 by auto-detecting JSON input.

If the a URL response, file contents or string provided as input contains only seralized JSON, it wraps the contents in a <script> tag and treats it like JSON LD embedded in a webpage.

This has some overhead, but ensures JSON parsing is subject to exactly the same flow as JSON LD embeded in pages.